requirements.txt
discord.py==1.5.1
pony==0.7.14
pycodestyle==2.6.0
psycopg2==2.8.6
python-dotenv==0.15.0
requests==2.25.0
